Delegate compression to servlet container
