+++ README.md
@@ -2,7 +2,6 @@
 * [Getting Started](docs/getting-started.adoc)
 * Developer Reference
   * [URL binding reference](docs/reference.adoc)
-  * [HTTP compression](docs/compression.adoc)
   * [Internationalization support](docs/i18n.adoc)
   * [Javadoc](https://javadoc.jenkins.io/component/stapler/)
   * [Stapler jelly tag library reference and XML Schema](docs/jelly-taglib-ref.adoc)
