+++ core/src/main/java/org/kohsuke/stapler/StaplerResponse.java
@@ -229,13 +229,18 @@ public interface StaplerResponse extends HttpServletResponse {
      *
      * @param req
      *      Used to determine whether the client supports compression
+     * @deprecated use {@link #getOutputStream}
      */
+    @Deprecated
     OutputStream getCompressedOutputStream(HttpServletRequest req) throws IOException;
 
     /**
      * Works like {@link #getCompressedOutputStream(HttpServletRequest)} but this
      * method is for {@link #getWriter()}.
+     *
+     * @deprecated use {@link #getWriter}
      */
+    @Deprecated
     Writer getCompressedWriter(HttpServletRequest req) throws IOException;
 
     /**
